+++ amplpy/__init__.py
@@ -12,7 +12,7 @@ from .set import Set
 from .parameter import Parameter
 from .entity import Entity
 from .dataframe import DataFrame
-from .utils import multidict
+from .utils import multidict, register_magics
 from .environment import Environment
 from .ampl import AMPL
 __version__ = '0.6.5b1'